CLN: avoid accessing private functions
